#134 rename and restructure
