+++ packages/web/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import LandingHome from './components/spa/landing_home_page';
+import LandingHome from './components/spa/MainPage';
 import LandingAdminPage from './components/admin_page/landing_admin_page';
 import Dashboard from './components/admin_page/admin_dashboard';
 import MemberActions from './components/admin_page/members_page/single_member';
