+++ packages/web/src/components/spa/MainPage.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import './main.css';
-import { Members } from './Members/Members';
+import { MembersSection } from './MembersSection/MembersSection';
 import { NavBar } from './Navigation/NavBar';
-import { Welcome } from './Welcome_section/Welcome';
+import { LandingSection } from './LandingSection/LandingSection';
 
 const LandingHome = () => {
     return (
         <div className="main">
             <NavBar />
-            <Welcome />
-            <Members />
+            <LandingSection />
+            <MembersSection />
         </div>
     );
 };
